+++ packages/spreadsheet-core/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./types";
 export * from "./utils/gridUtils";
+export * from "./utils/csv";
 export * from "./SheetAdapter";
+export * from "./InMemorySheetAdapter";